* Warning removal for Visual Studio's warning level 4.

+++ src/buffer.c
@@ -2,7 +2,7 @@
 #include <string.h>
 #include <malloc.h>
 
-void buffer_openspacehelper(BUFFER *buff, size_t amount)
+void buffer_openspacehelper(BUFFER *buff, int amount)
 {
     if (amount + buff->pos < 1024)
         buffer_resize(buff, 1024);
@@ -11,7 +11,7 @@ void buffer_openspacehelper(BUFFER *buff, size_t amount)
 }
 
 
-void buffer_resize(BUFFER* buff, size_t size)
+void buffer_resize(BUFFER* buff, int size)
 {
     if (size == 0)
     {
@@ -35,7 +35,7 @@ void buffer_resize(BUFFER* buff, size_t size)
     if (buff->buffer == (char*)&buff->static_buffer) {
 	buff->buffer = (char*)malloc(size);
 	memcpy(buff->buffer, &buff->static_buffer, buff->pos);
-    buff->buffsize = size;
+	buff->buffsize = size;
     } else {
         buff->buffsize = size > buff->buffsize * 2 ? size : buff->buffsize * 2;
         buff->buffer = (char*)realloc(buff->buffer, buff->buffsize);
